+++ hyrisecockpit/frontend/src/meta/workloads.ts
@@ -1,4 +1,5 @@
 import { Workload, WorkloadProjectionData } from "../types/workloads";
+import { isInTestMode } from "../../config";
 
 const workloadProjectionData: Record<Workload, WorkloadProjectionData> = {
   tpch01: {
@@ -34,4 +35,14 @@ export function getDisplayedFromTransferred(transferred: string): string {
   return Object.values(workloadProjectionData).find(
     (workload) => workload.transferred === transferred
   )!.displayed;
+}
+
+export function changeTableName(table: string): string {
+  if (isInTestMode) return table;
+
+  const words = table.split("_");
+  if (words.length === 4) {
+    return `${words[0]} (${words[1].toUpperCase()} SF ${words[2]}.${words[3]})`;
+  }
+  return `${words[0]} (${words[1].toUpperCase()} SF ${words[2]})`;
 }
